Move gba_*.h into subdir, add gba.h
