+++ src/clear_ram.c
@@ -1,6 +1,6 @@
 #include "memory.h"
 
-#include "gba_memory.h"
+#include "gba.h"
 
 void clear_ram(void) {
     bit_fill(3, 0x00, VRAM_BASE, VRAM_SIZE, 0x10);
