+++ src/funcs.h
@@ -1,7 +1,10 @@
 #ifndef FUNCS_H
 #define FUNCS_H
 
+#include "types.h"
+
 void clear_ram();
+void dma_transfer(int dma_channel, void *src, void *dst, u32 len, u8 bit_size);
 void init_sound();
 void load_intr_code();
 void read_sram();
