+++ src/gba.h
@@ -1,0 +1,15 @@
+#ifndef GBA_H
+#define GBA_H
+
+/* This just serves as an easier way to get all the GBA headers */
+
+#include "io.h"
+
+#include "gba/display.h"
+#include "gba/dma.h"
+#include "gba/interrupt.h"
+#include "gba/keys.h"
+#include "gba/memory.h"
+#include "gba/waitstate.h"
+
+#endif /* GBA_H */
