+++ src/gba/keys.h
@@ -1,7 +1,7 @@
 #ifndef GBA_KEYS_H
 #define GBA_KEYS_H
 
-#include "gba_memory.h"
+#include "gba/memory.h"
 #include "types.h"
 
 #define REG_KEY_INPUT (REG_BASE + 0x130)
