+++ src/init_game.c
@@ -1,12 +1,7 @@
 #include "callbacks.h"
 #include "funcs.h"
 #include "game_modes.h"
-#include "gba_display.h"
-#include "gba_dma.h"
-#include "gba_interrupt.h"
-#include "gba_keys.h"
-#include "gba_memory.h"
-#include "gba_waitstate.h"
+#include "gba.h"
 #include "globals.h"
 #include "io.h"
 #include "types.h"
